pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-rdf-mapper"
version = "0.2.0"
description = "Convert tabular data (CSV, Excel, JSON, XML) to RDF triples aligned with OWL ontologies using SKOS-based semantic mapping"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "RDFMap Contributors", email = "rxcthefirst@gmail.com"}
]
maintainers = [
    {name = "RDFMap Team", email = "rxcthefirst@gmail.com"}
]
keywords = [
    "rdf", "ontology", "semantic-web", "knowledge-graph", "owl", "shacl",
    "linked-data", "data-conversion", "skos", "owl2", "ttl", "json-ld",
    "csv-to-rdf", "excel-to-rdf", "json-to-rdf", "xml-to-rdf",
    "semantic-mapping", "ontology-alignment", "data-integration"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "rdflib>=7.0.0",
    "polars>=0.19.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyshacl>=0.25.0",
    "typer>=0.9.0",
    "PyYAML>=6.0.1",
    "python-dateutil>=2.8.2",
    "rich>=13.7.0",
    "click>=8.1.7",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/rxcthefirst/SemanticModelDataMapper"
Documentation = "https://github.com/rxcthefirst/SemanticModelDataMapper#readme"
Repository = "https://github.com/rxcthefirst/SemanticModelDataMapper"
Issues = "https://github.com/rxcthefirst/SemanticModelDataMapper/issues"
Changelog = "https://github.com/rxcthefirst/SemanticModelDataMapper/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.1",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "types-PyYAML>=6.0.12",
    "types-python-dateutil>=2.8.19",
    "psutil>=5.9.0",  # For memory profiling in benchmarks
]

[project.scripts]
rdfmap = "rdfmap.cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=rdfmap --cov-report=term-missing"
